Make values included in status configurable via cli
Let the feature-creeping begin!

Jokes aside, we'll probably want to include different things on
different machines. Still, we don't want to bother with short and long
options (at least for now) and a sequence of specifiers appears more
natural anyway.

In our case, this sequence will be a single string with each character
specifying one "thing" to include, in the order specified. Which
impliess that we can also alter the order. To some degree, that is:
date/time remains at the line's beginning.

"User-facing" documentation will follow at some point.