Add initial rust setup
`status` served its purpose when we tried out sway a while ago, but when
we considered some future extensions we were ultimately confronted with
limitations of the design in C. The one-to-one mapping of sources and
entries prevented aggregation of multiple sources for one entry. And
the inability to hold state meant we could not easily do estimations
based on "historic" data, e.g. for battery drainage.

We could revisit our design and fit all this in somehow, probably with
an unconfortable amount of `void*`. However, we decided to take this
opportunity to reimplement `status` in Rust. We'll go for the current
feature-set, but a more extensible design.

This change introduces the initial setup for a Rust binary, with
`anyhow` as a dependency because we are definitely too lazy for anything
else error handling wise. Future changes will extend it bit by bit.