main: basic setup for prodicing a line every second
... early in that second

One core functionality of the original `status` is printing one line
every second, early in that wallclock second. This change introduces the
basic setup to do just that. Currently, that line remains empty, but
future changes will allow for configurable contents.

We would have liked to do writing to stdout and waiting for the next
tick via the IO uring as well. Given the `IOSQE_IO_DRAIN` drain flag
it should be possible to only issue one set of submission queue events
per tick that "do it all". However, when we actually tried this we
observed some bad behaviour and absurdly huge IO loads. So for now, we
do these things the classic way. We may still go for "full IO uring" in
the future.

Another decision we made delibarately was to not write to stdout
directly but to fill a buffer first. This should safe us some repeated
allocations. And if we want to write those lines via the IO uring in the
future, we'll need to handle the output buffer ourselves anyway.