read: add abstraction for processing read Items
We recently introduced a type for handling and dispatching single,
potentially recurring reads. The idea, of course, is to use a set of
them with an io-uring. This involves submission and awaiting as well as
dispatching results.

This change introduces a type for handling these tasks. We decided to
encapsulate the `iou::IoUring`, since we depend on exclusive usage of
that io-uring anyway.