read: add abstraction for reads of a file into a buffer
... with subsequent processing.

This abstraction will help collecting values from various files in
`/sys` and `/proc` and processing the values afterwards. These can be
thought of as the value collection aspect of `struct status_entry` in
the C version. However, rather than a single `fn` we use a dynamically
dispatched trait, which allows storing arbitrary (structured and
non-`str`) results and othger state in the traits impl.

On the flip side, we let go of the capability of forwarding a "view"
into the buffer to the printing part we'll introduce eventually. We
solved the lifetime-puzzle in a previous attempt which had a slightly
different setup (it stored the extracted value in the `ReadItem`
directly). However, that had some short-comings and we gave up on making
it work for this version. Parsing numbers just to print them again feels
a bit iffy, but it's not like we'd never want to do anything else with
them.