main: add IO uring setup and usage logic
Because we obviously will be using an IO uring. Otherwise we would not
have written all this code tailored to it. And with these changes, we
are still not making actual use of the IO uring, as `reads` remains
empty. Future changes will put stuff in it. That's why we made it `mut`.