entry: add trait for status line entries
`status` is intended to print lines with an immutable set of entries
with optional labels for each of them and changing values. In the C
version, the values would be strings lifted from a buffer filled though
a read. In the Rust version, we intend to allow more diverse types of
values, including ones determined from inputs from multiple sources.

The representation of an entry needs to be flexible enough for these
types of values, and ideally we want to have some level of
composability. The latter means that the most simple solution of
dispatching on `Display` trait objects is not an option. Rather, we need
(or want) the possibility of fiddeling around with the thing we want to
display. For example, we want the possibility to control the precision
with which an `f32` is displayed.

So now we arrived at a trait for creating `Display`s of specific type,
which may nor may not be some sort of reference. The trait can be
implemented for data sources as well as adapters. We would havb liked to
put in in a `Box` and dispatch dynamically on some formatting fn, but
that isn't possible due to the `Display` type. Since we really only need
to dispatch on a single function, we decided to just move it into a
closure which we can then `Box` as a trait object.