deps: add libc as a direct dependency
Time-related stuff is hard. Even after making sure we fire up on exact
wallclock second ticks, we still need to get the actual localtime, and
format it.

There are high-level crates for that, namely `time`. Unfortunately, past
design explorations unveiled that `time` apparently doesn't like events
being in transit on an io-uring, and looking at its implementation it
definitely does a few things which strike us as unnecessary in our
situation.

So we're stuck with using `libc` directly for now. At least that works.