entry: add entry type for the current local time
The original `status` includes the current wallclock time in the status.
And because it was the only thing we didn't read from a file descriptor,
its the only thing which is not configurable in the C version.

But we develop the Rust version for the very reason to overcome the
limitations of the C design. And so, lo and behold, we can make the
local time an entry just like any other.