deps: add iou as a dependency
Like the original `status`, the Rust version will use IO uring. There
are some crates offering various levels of abstraction. We won't use
`async`, but do the dispatching ourselves. Unfortunately, some of the
non-async crates out there more or less assume a continuous dispatching
flow, while we specifically want to submit a packet of items and
dispatch them one-by-one, then continue with formatting. Thus, we want
something like `io_uring_wait_cqe`.

In the end, we decided for `iou`, which is an abstraction on top of the
same `libiouring` we were using in the C version. At the time of writing
this, the current release is a few years old, and there are a few minor
issues and pull requests open on the public repository. But it appears
to still be the most reasonable option. We're not going to write our own
abstraction right now...