source: add trait for value sources
The `status` tool ultimately just prints values it finds in `/sys` and
`/proc` files. In the C version, these values would be strings extracted
from those files, which happened to be numbers. While this approach was
fine for the few things we support in the C version, it doesn't work so
well if we want to perform calculations on those numbers, nor does it
allow for extraction of multiple/structured items from a single file
such as `/proc/meminfo`. Therefore, we decided to make the interface
typed in the Rust version.

This change introduces that interface in the form of a trait. Future
changes will introduce various impls, most of which will also impl
`read::BufProcessor`. In most cases those impls will provide simple,
scalar values for which we don't want the overhead of a reference.
However, we also need to accomodate for some values that are expensive
to clome. We thus decided to introduce an intermediate `Borrow`,
allowing us to return either owned data or a reference depending on the
needs. We'll utimately still need to handle a reference somewhere, as
`Borrow::borrow` will return one. However, having that on the caller's
site will mean we have concrete types, and the whole `Borrow` will be
optimized away. At least that's what we're hoping for...