source: add a source for a single word
The original C version would allow extracting the first value (or word)
from `/proc/loadavg`. Since we want to provide the same functionality in
the Rust version, we need a `Source` providing that functionality. And
because it takes virtually no additional effort and may come in handy
with other files, we make it more generic than the old version that
could only retrieve the value verbatim.