source: add extractor specifically for PSI info
The original C version would allow extracting 10min average pressure
info. Since we want to provide the same functionality in the Rust
version, we need a `Source` providing that functionality. Since the
average is provided as a rational number, we decided to parse it as a
floating point. An `f32` is sufficient since the numbers are not that
precise to begin with, and we don't do any calculations with them.