spec: add data structure for single spec
Narurally, we want the ability to configure the contents of a status
line. In the C version, we did this through a sequence of characters,
each of which would generate one or more entries. While this allowed for
a very concise configuration, the format had obvious limitations. For
example, a `p` would lead to the inclusion of all PSI entries, without
any means to select only a few of them. We considered some schemes for
subscripts, but the termination of subscripts would be a source of
dissatisfaction.

The limitation to single characters for selection of essentially an
entry class is also something we consider a hurdle for future
extensions. Short options are nice, but a very finite set forces us to
choose more and more obscure options.

We thus decided to not make the CLI compatible with the C version.
Instead each spec is defined through its own command line argument, and
we'll allow for longer (i.e. non-single-character) identifiers for entry
classes.

This change introduces a type representing such a specification, with a
`FromStr` impl and some utility functions. Since a `Spec` will be quite
short-lived, we decided to have it reference parts of the string we
parsed it from rather than having it own the data.