spec: add utility for preventing creation of duplicate read items
For a number of reasons, we may end up in a situation where multiple
entries make use of information extracted from the same file. Ideally,
we want to use the same `source::Source`, and by extension the same
`read::Item`.

We are pretty sure that we'll never use `Source`s of different types in
`read::Item`s targeting the same file. And even if we did, we'd likely
want to use a `Source` which features all the values we want instead of
two distinct `Source`s.

This change introduces a utility which lets us transparently construct a
`Source` (or rather, a `BufProcessor`) and register a `read::Item` to
feed it, or re-use an existing `Source` if we already have one for the
same path.

The downside is that this will likely contribute to memory fragmentation
during startup.