spec: add fn for applying a spec
We recently introduced the `Spec` type for parsing and processing entry
specifications. What we were lacking was code deciding what kind of
entries to create for a given spec. This change introduces a fn doing
just that. And it's still private to the `spec` module. We'll introduce
a public one for processing multiple specs from command line arguments
in the future.