entry: impl Entry for read::Ref<source::Source>
The Rust version was motivated partially by more complex entries based
on multiple sources. However, we still have lots of entries based on a
single source. Often times, that `Source` will provide a trivial value
that we can, more or less, display as is.

This change makes this easy, at least for most cases, by implementing
`Entry` for `read::Ref` to a `Source`.