entry: allow labeling Entrys
Most status line entries will display numerical data. Naturally, we want
to have labels stating what kind of data we are viewing. In the C
version, every entry had a label, but with entries like the date-time
one, we already have at least one for which we don't want a label. We
could well make the label optional, but this scheme allows more
use-cases.