entry: add utility for formatting a list of entries
We previously decided to keep and dispatch status line entries as
`dyn Fn(&mut fmt::Formatter)`. These are not `Display` and neither are
there currently any (stable) utilities (in `std`) for making them
`Display`. And implementing those traits is really the only (sane) way
to get our hands on a `fmt::Formatter`.

We could have gone with a `dyn Fn(&mut Write)` instead, with a specific
`Write`. However, given that we want to format a space-separated list
anyway, we can simply roll a slightly more specialized version of a
general list formatting utility and use its `fmt::Formatter`.