Cargo: add a custom release profile optimizing for size
`status` is a long-running service which will be active regularly, but
only for a very short time period. Getting code (and data) into caches
or not having them evicted has probably more effect than anything else.
We therefore decide to go for as small a binary as reasonably possible.
And because it will be small, we may as well turn on LTO.

Besides, we like small binaries.