README: rewrite relevant parts for the Rust version
We rewrote `status` in Rust, now we should follow up on the
documentation (outside of the code), which only consists of the
`README.md`.

Im many cases we only needed to adjust a few formulations. However,
aside from the "Build" section we also had to completely rewrite the
"Usage" section sicne the CLI is different for the Rust version.

We also decided to rewrite the introduction of the "Motivation" section
and rename it "Background" because of all the watrer that went under the
bridge since we wrote it.