read: add fn for creating a Ring using a specific io_uring::Builder
Previously, the only way to create a `Ring` was using its `TryFrom`
impl. We did so because we really, really wanted to decide the number of
entries for the io-uring in code associated with `Ring` rather than have
(documented?) requirements on the ring provided from the outside.

This, however, also meant we couldn't provide any (alternative) setup
flags. That we can achieve by handing in some kind of builder for an
uring instance. Fortunately, `io-uring` does provide a builder.

This new fn allows handing in a builder.