meminfo: add dedicated module for meminfo related stuff
Handling `/proc/meminfo` does take some amount of code that we are
unconfortable with putting in `source` or `read`. And it all
conceptually belongs together. We threfore decided to put it all in a
dedicated module.

This change introduces this module with the container for all the things
from `/proc/meminfo` we may be interested in. Future changes will make
this usable in `Entry`s and introduce entry specifications.