source: add trait for updateable Sources
With our current design, data `Source`s are usually `BufProcessor`s
which extract a value and store it internally. This was good enough for
the current functionality, but it also meant some code duplication
w.r.t. the value storage.

In the future, we may want to do some form of aggretation inside a
source, and for those we want to avoid duplicating the extraction part.
The common strategy to avoid duplication in this type of scenario is
composition.

In order to enable composition, this change introduces a new trait for
updateable `Source`s. We will implement this for types holding values.
In the future, `BufProcessor`s will use this trait to push values into
`Updateable` `Source`s.