source: impl Source for Option
We intend to make `Source`s composable, at least from an extraction
and an accululation/value storage stage. All of our current `Source`s
store their internal value in an `Option`, i.e. the accumulation stage
we definitely need is something equivalent. So why not just use a bare
`Option`?

This change impls `Source` for `Option`, and a future change will impl
`Updateable` for `Option`. This allows that type to be wrapped by future
`BufProcessor`s which forward access to this `Source` via their `Source`
impl.