read: add BufProcessor for extracting a Word
`source::Word` is a `Source` which extracts the first word from a file,
parses and served it via the `Source` trait. This was more than
sufficient for the current use-case of extracting the system load from
`/prog/loadavg`. However, we intend to introduce functionality which
require some form of accululation of "first words in a file", and we
don't want to duplicate the extraction logic for all forms of
accumulation.

We therefore decided to allow for the composition of `Source`s, and
introduced changes for reaching that goal. We also deemed that the most
suitable place for `BufProcessor` `Source`s was the `read` module after
all. We'll therefore make `read` depend on `source`, and remove any
dependency from the latter on the former.

This change introduces a version of `Word` that pushes the parsed value
into an `Updateable`, presumably a `Source`. Migration of users and
removal of `source::Word` will be done by a future change, since we
didn't just move `Word` but also changed the type's definition.

And since we're at it, we also added some tests.