meminfo: impl read::BufProcessor for MemInfo
We intend to support displaying some values from `/proc/meminfo` as part
of the status line. For that purpose, we already introduced a container
for those values in a dedicated module which is also meant to hold any
impls associated with that type.

And because this is kind-of special, with a very narrow use-case, we
already implemented `Source` for `MemInfo` directly. And the same
rationale applies to `BufProcessor`.