meminfo: impl Source for MemInfo
We intend to support displaying some values from `/proc/meminfo` as part
of the status line. For that purpose, we already introduced a container
for those values in a dedicated module which is also meant to hold any
impls associated with that type.

And because this is kind-of special, with a very narrow use-case, there
is not really any point in introducing a new type serving as a `Source`
when it only holds an already very specific `MemInfo`.

We therefore decided to impl `Source` for that type directly.