scale: add trait for scales
In the future, we may want to allow for entries with values which may
potentially cover a large range, such as memory sizes. Instead of
displaying an unreasonably large number of, say, kB, we want to display
a number in a larger unit depending on the actual value. And we may not
only want this for units with SI prefixes, but also for the "binary"
scale (kiB, MiB, GiB, ...) and times (so seconds, minutes, hours, ...).

This change introduces a trait for enums expressing these kinds of
scales, allowing us to abstract over them. Future changes will introduce
implementers and an entry type capable of doing the scaling.