scale: add type for scaled values
In the future, we may want to allow for entries with values which may
potentially cover a large range, such as memory sizes. Instead of
displaying an unreasonably large number of, say, kB, we want to display
a number in a larger unit depending on the actual value. And we may not
only want this for units with SI prefixes, but also for the "binary"
scale (kiB, MiB, GiB, ...) and times (so seconds, minutes, hours, ...).

For accomodating such use-cases, we recently introduced the `Scale`
trait for handling different scales. We did, however, not yet have any
means to represent a value assiciated with a scale. This change
introduces such a type, with a function for choosing a scale convenient
for displaying.