scale: add binary (kibi, Mebi, Gibi, Tebi) scale
In the future, we may want to allow for entries with values which may
potentially cover a large range, such as memory sizes. In order to
display them in a more human readable format, we recently introduced the
`Scaled` type for values and the `Scale` trait for scales.

This change introduces a binary scale usable for memory sizes. Future
changes may introduce additional scales for other purposes.