entry: add an Entry which automatically scales its Display type
potentially cover a large range, such as memory sizes. In order to
display them in a more human readable format, we recently introduced the
`Scaled` type for values and the `Scale` trait for scales. What comes
out of a `Source` will still be a "raw" value which needs to be
associated with a `Scale` afterwards at some point. Which means we need
an `Entry` type/adapter for that.

This change introduces such an `Entry`, using the existing `Scaled` as
its `Display` type. In addition to attaching a scale it also performs
autoscaling.

Future changes will add fns for creating such an entry and tests.