entry: add an Entry mapping the value from a Source
In the future, we may need to apply some sort of transformation on a
non-`Display` value retrieved from a `Source` for an `Entry`, e.g. if
the source provides a complex value such as a struct holding data from
`/proc/meminfo`. This change introduces an `Entry` type for that.

We would have liked to use the existing `Entry` impl on
`read::Ref<source::Source>` and introduce another `fn` to `Entry` for
the mapping. However, that conflicted with our (reasonable) requirement
that an `Entry`'s `Display` type does, in fact, implements `Display`.