deps: add enum-map as a dependency
We plan to introduce a data `Source` for values from `/proc/meminfo`.
That file exposes values in the form of a list of labeld numeric values,
which we consider all being of the same type.

We already want an enum for sub-specs and `Entry` labels. If we can just
have some form of map from that enum to a value, that lets us parse
those values with less of a headache, and also makes retrieval
straight-forward.

`enum-map` lets us make that map an array.