entry: add Entry with a unit of measurement
Currently, we only support displaying plain unitless values. However, we
intend to support displaying various values which do have some unit
attached to them. These include (memory) sizes and temperatures. And
even for percentage we'll want to include the percent-sign as a "unit".

This change introduces an `Entry` adapter with allows attaching a unit.
Future changes will introduce a fn for creating such an `Entry` as well
as tests.