scale: add scale for time durations
We plan on supporting some information on the current battery state,
including an estimate on battery lifetime. And we want that time to be
readable.

Note that among scales, this is a bit of an oddball, both because the
scales are actually units and because the factors from one scale to the
next are not uniform. In fact, the latter motivated some design choice
for the `Scale` trait.