entry: impl Entry for Fn() -> Option<Display>
Currently, almost all the `Entry`s we are using are fed from a single
`Source` without any transformations happening in between. However,
features we plan to implement in the future require some transformation,
be it in order to make the value a `Source` provides `Display` or
combine multiple `Source`s.

For a subset of these use-cases, we already introduced a `Mapped`
`Entry` type in

	e0994ed4d3eb919e690a4dd6b7beb403a79b003a

However, it turned out that the use-case `Mapped` covers to be more of a
niche, and its design was too inflexible for reasonable extension. In
particular, it does not allow for combination of values from different
`Source`s, which we may not only need to do in a "zipping" fashion but
also in a "selective" manner --which means we'll impl `Entry` for
`either::Either` at some point.

So we decided to go with something different and _just_ implement
`Entry` for a subset of `Fn`, already. We tried to not do this, because
it's all too tempting to replace all the specific `Entry` types with
closures. And maybe that's what we should do at some point after all.

Future changes will introduce free convenience fns for creating `Fn`s
functioning as `Entry`s. After we introduced a replacement, we'll also
get rid of the existing `Mapped` which we did not (yet) use anyway.