entry: add fn for creating an Entry mapping a source
In the future, we may need to apply some sort of transformation on a
non-`Display` value retrieved from a `Source` for an `Entry`, e.g. if
the source provides a complex value such as a struct holding data from
`/proc/meminfo`. This change introduces a function for creating an
`Entry` from a `Source` and an arbitrary mapping. This fn serves as a
replacement for the `Mapped` `Entry` introduced in

	e0994ed4d3eb919e690a4dd6b7beb403a79b003a

A future change will remove `Mapped`.