entry: add fn for creating an Entry zipping two sources
In the future, we'll likely need to combine values from multiple
`Source`s. The most basic, and probably a common, version of that is
zipping two values, e.g. in order to calculate a percentage.

This change introduces a utility for creating such an `Entry` from two
`Source`s and a zipping fn.