power: add type representing a power supply, in dedicated module
We plan to support inclusion of battery related information. Batteries
are a kind of power supply, and those expose values through the sysfs.
Each device of the power supply class is represented as one directory in
`/sys/class/power_supply`. Each of those directories in turn contains
files, each of which exposes a single value.

Thus, for all the information we want for some battery, we'll need to
open multiple files, and we want to do that in a way that both doesn't
require loads of boiler plate in the `spec` module and is at least
somewhat resilient against directory movement and unlinking. We thus
decided to introduce a new type encapsulating an `openat::Dir` and,
since we'll need it for labels, a name of the device.

We make the new type represent all power supplies. Future changes will
introduce fns for retrieving the kind of power supply and fns for
opening specific files corresponding to values we're interested in.
Since the type will grow and require some accilary enumerations, we
decided to put it into a dedicated `power` module.