deps: add openat as a direct dependency
All files we're opening currently have a single, well defined location,
that is, a path, in `/proc`. However, some of the values we plan to
source for entries in status lines are associated with some device and
exposed somewhere in `/sys`, usually under some directory associated
with that device. In addition, we may need to combine values from
multiple files associated to one specific device.

Theoretically, the directory representing the interface for some device
may move, or get unlinked. This means that we may get into trouble when
trying to open multiple files for one device if we're relying on path
stability. Therefore, we'd like to use `openat(2)`. Curiously, `std`
apparently doesn't provide any way to make use of it. We thus have to
use solutions outside `std`.

The current go-to solution appears to be the `openat` crate. Or rather,
a fork of it that accumulates patches since the original currently
doesn't appear to receive any maintenance. We decided to use the
original `openat` anyway. Everything we need is there and there appears
to be no open bug which would hurt us. And since we'll use it "only"
for opening `std::fs::File` (and that's an `std` type), we can replace
it more easily than we did `iou` back then, should that become
necessary.

Maybe this functionality will make it into `std` one day, anyway, as it
ought to. And we're still hoping that `openat` receives maintenance
again after the war, whenever that will be.