power: add fns for retrieving charge information of a battery
The most basic information we're interested w.r.t. batteries is the
state of charge, i.e. how "full" or "empty" it is. We are usually
interested in a percentage which is easy to grasp at a quick glance.
We need to calculate that percentage ourselves from values exposes via
sysfs.

Power supplies of type "Battery" expose the current charge and the
charge when the battery is full or empty, and for those both the design
and current value. Supposedly, the latter may decrease as the battery
ages. This change introduces fns for opening those files. We chose to
omit fns for design values, since we don't consider them immediately
relevant for users.