spec: use device and inode nums as keys for map in ReadItemInstaller
The `ReadItemInstaller` exists because we want to re-use `read::Item`s
when possible. And we'd do that by storing them in a hash map with the
path of the file we're reading as a key.

This obvious but naïve solution breaks down if we don't have the path,
e.g. because we're using `openat` and don't manually track paths. And
you use `openat` because files could move. Sure, `/proc/loadavg` will
not move, but we may well end up sourcing information from some devices
which are plug-and-play enough that we'd better not depend on the path.

On Unixes, we can (usually) use the file's device and inode number as a
unique identifier. Inode numbers may be re-used, but only if the file
vanishes. And well, we do have the file opened, so... it will not. Thus,
we decided to switch to using those as keys for our map.

Note that this switch doesn't allow for race-free installation of
`read::Item`s for files that may get moved or unlinked. Fns allowing for
that will be added by some future change.