spec: add fns for installing with Files rather than Paths
The `ReadItemInstaller` exists because we want to re-use `read::Item`s
when possible. Previously, the only possibility to install new items was
by also supplying an `std::path::Path`. However, we plan to introduce
entries, eventually, for which we won't have a path. For example, we may
opening files using `openat` in order to avoid races.

Still, we're pretty sure we'll always have an `std::fs::File`, as we
need one to construct a `read::Item`. Thus, we decided to introduce fns
to `ReadItemInstaller` which allow installation of `read::Item`s by
`std::fs::File` rather than `std::path::Path`. Because the latter are
still convenient for files at guranteed static locations such as
`/proc/loadavg`, we keep them.

Due to this change, the fns for path-based installation open the files
early rather than doing a `stat`. There's pros and cons to both
solutions, but we chose this one because it required the least amount of
code.