power: add fn for getting all power supplies
We already have introduced means to create a `Supply` by name (which
will error out if that supply does not exist). However, if a user did
not specify a supply directly, we'll want to go through them all and
select some.

This change introduces a fn which creates an `Iterator` of `Supply`s
that are constructed directly from a directory listing.