entry: add FormatterFn utility for formatting via Fns
When introducing `Entry`, we said that returning a boxed `Fn` from
`Entry::into_fmt` was good enough. That we wouldn't necessarily need
some `fmt::Display`.

But having a `fmt::Display` does have its merits. If `EntriesDisplay`
would accept `Box<Display>`s, we could just push in `String`s, and in a
way that would not be as cashe-unfriendly as if we stuffed it into a
`Fn`. This would allow for "free" labels. We could also completely
replace `Labeled` `Entry`s.

Thus, we decided to eat our hat. But we still don't want to introduce a
dedicated type tailored for `Entry`s specifically. What we want instead
is a boxed `std::fmt::FormatterFn`, but `FormatterFn` is not stable yet.
We therefore bring our own to use in the meantime.