power: add fn for retrieving the current flowing out of a battery
We could easily estimate the remaining time until the battery runs empty
by observing the `charge_now` value and calculating the rate by which it
decreases. However, we sometimes also get a value telling us the current
directly. There may also be batteries exposing a `charge_avg`, which
would make our lives even easier. However, we've yet to encounter
hardware featuring that.

This change adds a fn for retrieving the file for that value.