spec: add battery status to spec
The percentage of charge the battery holds is useful. However, the
battery status and time remeining until the battery runs dry is even
more important in some situations.

We decided to combine both, because we can imply the discharging state
easily by displaying our estimate of the remaining battery life.
However, this makes this also the most complicated `Entry`.

We also always use `current_now` for now, although we know that it is
definitely not present for every battery out there. And it may flutuate
too much between sampling points, leading to M$-copy-dialoge grade
estimates even with the moving average. We'll see.