read: add new "simple" BufProcessor
We have a few `BufProcessor`s, but the only one which can be considered
somewhat general purpose is `Word`. That processor splits the buffer at
ASCII whitespace, parses the first non-empty one and forwards the
resulting value to the underlying `Updateable` `Source`. This works well
for values that are "words", but we may well run into situations where
the value does contain some whitespace, and what we really need to
consider is lines rather than individual words.

We could introduce a `BufProcessor` specifically for lines, but that
would be almost a carbon copy of `Word`. The only differences would be
the name and the predicate used for splitting the buffer. So, the
obvious solution is have a `BufProcessor` that allows choosing the
splitting function.

This change introduces such a `BufProcessor` named `Simple`. Future
changes will migrate users of `Word` and remove `Word`.

`Simple` encapsulates the splitting function as a function pointer. We
did consider a making it generic over this value, i.e. having the
function pointer be a type parameter, but that's not possible with
current Rust (we compile with 1.77.1, 1.79.0 was released two days ago).
We could have also cooked up some trait, possibly only transporting a
`fn(&u8) -> bool` as a `const`. But we decided that we're of with
runtime variability here. It likely means less binary code, anyway.