read: remove now unused Word BufProcessor
`Word` splits the buffer it receives at ASCII whitespace, parses the
first non-empty one and forwards the resulting value to the underlying
`Updateable` `Source`. This works well for values that are "words", but
we may well run into situations where the value does contain some
whitespace, and what we really need to consider is lines rather than
individual words.

For this reason, we recently introduced a new `Simple` `BufProcessor`
with the plan to make it replace the existing `Word`. We also migrated
any users of `Word` over to `Simple`. With all the users gone, we can
now get rid of `Word`.