spec: import entry::Entry at top level of spec module
The `apply` fn is responsible for applying a specification, i.e. to
install `ReadItem`s and setting up `Entry`s. Thus, this fn grows with
each new specification we introduce, and those new parts start off with
a certain indentation level, making code formatting awkward. We thus
decided to move parts dealing with specific specifications out of
`apply`, and merely have it dispatch.

This change prepares this code-movement by lifting the import of
`entry::Entry` to module level, where it will be effective for all new
fns we introduce.