spec: migrate users of Entry::with_label to new free Label
Currently, we explicitely attach labels to `Entry`s, by wrapping them in
a `Labeled` `Entry`. The fn `Entry::with_label` makes wrapping easy and
convenient. However, we will liekly run into situations where a label is
associated with multiple `Entry`s, and while we could easily just attach
the label to the first of those, this doesn't really sit right with us.

For these situations, we recently introduced `entry::label`, a fn which
returns a `Display` with the input suffixed with a `:`. We plan to have
this mechanism replace `Labeled` `Entry`s. This change prepares for that
removal by migrating all users.