read: add fn for retrieving the raw file descriptor for a read::Item
The IO uring documentation advertises the possiblity to register files,
and states that registering files reduces per-IO overhead. Since we use
a fixed, unchanging set of files from which we read often, this could
have a positive impact on resource usage.

In order to register a file descriptor, we need to retrieve them from
the `Item` it is encapsulated in. This change introduces a fn for that.