read: register the file descriptor of all Items in Ring::new
The IO uring documentation advertises the possiblity to register files,
and states that registering files reduces per-IO overhead. Since we use
a fixed, unchanging set of files from which we read often, this could
have a positive impact on resource usage.

The obvious place to do this is in `Ring::new`, since it also "fixes"
the set of `Item`s and creates the IO uring.