source: make constant capping update rate a named constant
This is one of the cases where we want to document what a constant does.
Making it a named constant that we can slap a doc comment on obviously
helps, in addition to the obvious code-clearity a named constant brings
with it. In this case, we may also likely use this constant for other
purposes in the future.