source: add trait for signalling whether any processing is required
We currently poll every value we use in every main loop iteration, which
means we usually poll them every second. While this may be just a bit
eager for things like PSI and load, which change every other second, it
really is a waste of CPU time for some other values, most notably a
battery's `charge_full` value. And while making a decision does require
some amount of work in itself, it will most certainly not compare to any
system interaction we need to do, or parsing any value we read.

This trait introduces a trait which will let us abstract the decision
whether we need to do some work or not. Future changes will impl this
for `Source`s, and for `read::Item`. Based on that we can then _not_
prepare a given `read::Item` in `read::Ring::prepare`.