read: impl source::WantsProcessing for Simple
We recently introduced `source::WantsProcessing` for communicating
whether an entity, primarily a `Source`, will want to be driven. We
decided, nearly on a whim, that we'd just re-use the trait as the
interface through which a `BufProcessor` signals the same to the
`Item`.

A `Simple` `BufProcessor` really just needs to ask their `Source`.