read: impl source::WantsProcessing for PSI
We recently introduced `source::WantsProcessing` for communicating
whether an entity, primarily a `Source`, will want to be driven. We
decided, nearly on a whim, that we'd just re-use the trait as the
interface through which a `BufProcessor` signals the same to the
`Item`.

This change impls `source::WantsProcessing` for PSI. And since that
infromation changes every two seconds, we decided to just have it
updated once every second.