meminfo: impl WantsProcessing for MemInfo
We currently poll every value we use in every main loop iteration, which
means we usually poll them every second. Since this wastes a lot of CPU
time for items that don't have much potential to change rapidly, we plan
to fetch them less often. For signalling whether an item needs
processing with new values, we recently introduced `WantsProcessing`.
This change implements this trait for `MemInfo`. And because we can
allocate and deallocate huge amounts of memory in no time, we decided to
make this one of the things that better always receive an update.