source: add Source that should be updates with a "lower" rate
We currently poll every value we use in every main loop iteration, which
means we usually poll them every second. For some values, most notably a
battery's `charge_full` value, this is a waste of CPU time. For
communicating the updating requirement, we recently introduced the
`WantsProcessing` trait. However, the impl for the currently most used
`Source`, `Option`, will always demand an update as it doesn't offer any
means to store the data we'd need for a smarter decision.

This change introduces a new type which does impl `WantsProcessing`
along `Source` and `Updateable` and behaves just like `Option` w.r.t.
the latter two traits. We could have gone with an adapter of some sort,
but handling of the `Instant` the `Updateable` received an update would
be awkward. And, as future changes will demonstrate, we already have
some idea on how a `WantsProcessing` would look like for most, if not
all, other `Source`s.